+++ R/zzz.R
@@ -1,0 +1,4 @@
+.onLoad <- function(...) {
+  options(ggtrace.suppressMessages = FALSE)
+  options(ggtrace.as_tibble = FALSE)
+}